requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
numpy==2.4.2
opencv-python==4.13.0.90
Pillow==10.4.0
pytz==2024.2
SQLAlchemy==2.0.34
Werkzeug==3.1.3
face-recognition==1.3.0
face-recognition-models
setuptools
psycopg2-binary==2.9.10
